backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.32.0,<1.0.0
google-generativeai>=0.8.0,<1.0.0
yt-dlp>=2024.12.0
pydantic>=2.10.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
supabase>=2.11.0,<3.0.0
tenacity>=9.0.0,<10.0.0
